pyproject.toml
[project]
name = "fastapi-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "PyJWT>=2.9.0",
    "PyYAML>=6.0.0",
    "pydantic>=2.12.5",
    "psycopg[binary]>=3.2",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.40.0",
    "alembic>=1.17.2",
]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
pythonpath = ["src"]
cache_dir = "output/.pytest_cache"
